+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Add depth of field (dof) postprocessing effect
 - Add `SbNcbrTunnels` extension for for visualizing tunnels in molecular structures from ChannelsDB (more info in [tunnels.md](./docs/docs/extensions/tunnels.md))
 
-## [v4.2.0] - 2023-04-05
+## [v4.2.0] - 2023-05-04
 
 - Add emissive material support
 - Add bloom post-processing